+++ docs/CHANGELOG.rst
@@ -12,6 +12,8 @@ Unreleased
 Added
 -----
 - Add filtered BAM output to ``macs2-callpeak`` process
+- Add an option to use filtered BAM files from ``macs2-callpeak`` to
+  ``rose2``, ``workflow-macs-rose``, and ``macs2-rose2-batch``
 
 Changed
 -------